--- a/graph/mongo/quadstore.go
+++ b/graph/mongo/quadstore.go
@@ -380,7 +380,7 @@ func (qs *QuadStore) Type() string {
 	return QuadStoreType
 }
 
-func (qs *QuadStore) getSize(collection string, constraint *bson.M) (int64, error) {
+func (qs *QuadStore) getSize(collection string, constraint bson.M) (int64, error) {
 	var size int
 	var err error
 	bytes, err := bson.Marshal(constraint)