This is actually a more general optimization for a lot of backends, I'm just implementing it on Mongo first because it'll be easy to measure.

Mongo is also doing many size-fetching roundtrips which are unnecessary and slowing down baseline query speed a lot. So that needs to be fixed too.

The optimization is this: currently, while the indexes are built with prefixes, we're not taking advantage of that fact. This reduces the size of a lot of potential sets and speeds up queries a lot as a result.

This is a fairly naive implementation of the similar concept for Mongo. 

Benchmarks are these:
```
$ benchx mongo_indexed.benchstart mongobench.post
benchmark                                   old ns/op      new ns/op      delta	mult
BenchmarkNamePredicate                      25507744       24554295       -3.74%	-1.04x
BenchmarkLargeSetsNoIntersection            915498593      1435804572     +56.83%	+1.57x
BenchmarkVeryLargeSetsSmallIntersection     2224808408     1135664127     -48.95%	-1.96x
BenchmarkHelplessContainsChecker            1011967097     6988584246     +590.59%	+6.91x
BenchmarkHelplessNotContainsFilms           1944054542     175670016      -90.96%	-11.06x
BenchmarkHelplessNotContainsActors          1612441213     1005749592     -37.63%	-1.60x
BenchmarkNetAndSpeed                        817045039      568708805      -30.39%	-1.44x
BenchmarkKeanuAndNet                        521375267      391474443      -24.92%	-1.33x
BenchmarkKeanuAndSpeed                      486171401      369516419      -23.99%	-1.32x
BenchmarkKeanuOther                         1340610433     720032174      -46.29%	-1.86x
BenchmarkKeanuBullockOther                  2191868770     1031652926     -52.93%	-2.12x
```

Before this gets committed, I need to investigate HelplessContainsChecker and it's compatriot HelplessNotContainsChecker -- we've hit a really good optimization there, and, inversely, a large slowdown. Arbitrage opportunity :)