@barakmich 2015-04-19
Now we're cooking:
```
$ benchx mongo_indexed.benchstart mongo_indexed.bench3
benchmark                                   old ns/op      new ns/op     delta	mult
BenchmarkNamePredicate                      25507744       1348691       -94.71%	-18.90x
BenchmarkLargeSetsNoIntersection            915498593      222809540     -75.66%	-4.11x
BenchmarkVeryLargeSetsSmallIntersection     2224808408     411111060     -81.52%	-5.41x
BenchmarkHelplessContainsChecker            1011967097     578179240     -42.87%	-1.75x
BenchmarkHelplessNotContainsFilms           1944054542     6552483       -99.66%	-294.12x
BenchmarkHelplessNotContainsActors          1612441213     993648764     -38.38%	-1.62x
BenchmarkNetAndSpeed                        817045039      74003050      -90.94%	-11.04x
BenchmarkKeanuAndNet                        521375267      28129083      -94.60%	-18.52x
BenchmarkKeanuAndSpeed                      486171401      29718736      -93.89%	-16.37x
BenchmarkKeanuOther                         1340610433     100140074     -92.53%	-13.39x
BenchmarkKeanuBullockOther                  2191868770     167004309     -92.38%	-13.12x
```

Ready for review.