@kortschak 2015-04-20
LGTM with minor nits.

Should the fmt.Println("wat") be panic("wat")?