@barakmich 2015-04-25
The 'wat' was errant debugging. 

s/firstmongo/mongoIt (because it's important we're finding a mongo one)